[WIP] more contexts, this time around slow indexing and downloads
