+++ lib/pinchflat/fast_indexing/media_indexing_worker.ex
@@ -1,5 +1,7 @@
-defmodule Pinchflat.Workers.MediaIndexingWorker do
+defmodule Pinchflat.FastIndexing.MediaIndexingWorker do
   @moduledoc false
+
+  # TODO: make a startup task to rename all existing workers so they still run
 
   use Oban.Worker,
     queue: :media_indexing,
