+++ test/pinchflat/boot/startup_tasks_test.exs
@@ -1,4 +1,4 @@
-defmodule Pinchflat.StartupTasksTest do
+defmodule Pinchflat.Boot.StartupTasksTest do
   use Pinchflat.DataCase
 
   alias Pinchflat.Settings
